MoneyOps/voice-service/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
# Testing
pytest
pytest-asyncio
# LedgerTalk Voice Service Dependencies
# Python 3.11+

# LiveKit
livekit==0.12.0
livekit-agents==0.8.0
livekit-plugins-silero==0.6.0
livekit-plugins-groq

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# Utilities
tenacity==8.2.3

# API (optional health endpoint)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0